# poolmind/orchestrator/openapi.json
{"openapi":"3.0.0","info":{"title":"PoolMind Orchestrator API","version":"1.0.0","description":"The central orchestrator for the PoolMind pooled crypto arbitrage fund platform.","contact":{"name":"PoolMind Support","email":"support@poolmind.com"}},"servers":[{"url":"http://localhost:3000","description":"Development server"}],"components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT","description":"Enter JWT token"}},"schemas":{"TelegramAuthRequest":{"type":"object","required":["chatId","user"],"properties":{"chatId":{"type":"string","description":"Telegram chat ID","example":"279058397"},"user":{"type":"object","required":["id","username"],"properties":{"id":{"type":"number","description":"Telegram user ID","example":279058397},"username":{"type":"string","description":"Telegram username","example":"johndoe"},"first_name":{"type":"string","description":"User first name","example":"John"},"last_name":{"type":"string","description":"User last name","example":"Doe"},"language_code":{"type":"string","description":"User language code","example":"en"}}},"auth_date":{"type":"number","description":"Authentication timestamp","example":1662771648},"hash":{"type":"string","description":"Authentication hash for verification","example":"c501b71e775f74ce10e377dea85a7ea24ecd640b223ea86dfe453e0eaed2e2b2"}}},"UpdateProfileRequest":{"type":"object","properties":{"firstName":{"type":"string","minLength":1,"maxLength":50,"description":"User first name","example":"John"},"lastName":{"type":"string","minLength":1,"maxLength":50,"description":"User last name","example":"Doe"},"preferences":{"type":"object","properties":{"notifications":{"type":"boolean","description":"Enable notifications","example":true},"language":{"type":"string","minLength":2,"maxLength":5,"description":"Language preference","example":"en"}}}}},"LinkWalletRequest":{"type":"object","required":["walletAddress"],"properties":{"walletAddress":{"type":"string","description":"Stacks wallet address","example":"SP3FBR2AGK5H9QBDH3EEN6DF8EK8JY7RX8QJ5SVTE"},"signature":{"type":"string","description":"Wallet signature for verification (optional)","example":"0x1234567890abcdef..."}}},"User":{"type":"object","properties":{"_id":{"type":"string","description":"User ID","example":"507f1f77bcf86cd799439011"},"telegramId":{"type":"string","description":"Telegram user ID","example":"279058397"},"username":{"type":"string","description":"Telegram username","example":"johndoe"},"firstName":{"type":"string","description":"User's first name","example":"John"},"lastName":{"type":"string","description":"User's last name","example":"Doe"},"walletAddress":{"type":"string","description":"Linked Stacks wallet address","example":"SP3FBR2AGK5H9QBDH3EEN6DF8EK8JY7RX8QJ5SVTE"},"kycStatus":{"type":"string","enum":["pending","approved","rejected"],"description":"KYC verification status","example":"pending"},"isActive":{"type":"boolean","description":"Account active status","example":true},"preferences":{"type":"object","properties":{"notifications":{"type":"boolean","example":true},"language":{"type":"string","example":"en"}}},"createdAt":{"type":"string","format":"date-time","description":"Account creation date","example":"2023-01-01T00:00:00.000Z"},"updatedAt":{"type":"string","format":"date-time","description":"Last update date","example":"2023-01-01T00:00:00.000Z"}},"required":["telegramId","username"]},"AuthResponse":{"type":"object","properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","example":"User authenticated successfully"},"data":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/User"},"token":{"type":"string","description":"JWT access token","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."},"isNewUser":{"type":"boolean","description":"Whether this is a new user registration","example":false}}}}},"SuccessResponse":{"type":"object","properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","example":"Operation completed successfully"},"data":{"type":"object","description":"Response data (varies by endpoint)"}}},"ErrorResponse":{"type":"object","properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Error message"},"errors":{"type":"array","items":{"type":"string"},"description":"Detailed error messages","example":["Field validation failed","Required field missing"]}}}},"responses":{"UnauthorizedError":{"description":"Access token is missing or invalid","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"},"example":{"success":false,"message":"Access token required"}}}},"ValidationError":{"description":"Validation failed","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"},"example":{"success":false,"message":"Validation failed","errors":["initData: Telegram initData is required"]}}}},"NotFoundError":{"description":"Resource not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"},"example":{"success":false,"message":"User not found"}}}},"ServerError":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"},"example":{"success":false,"message":"Internal server error"}}}}}},"tags":[{"name":"Authentication","description":"User authentication and profile management"},{"name":"Balance","description":"Token and wallet balance queries"},{"name":"Pool Operations","description":"Pool deposits, withdrawals, and balance queries"},{"name":"Admin","description":"Administrative operations"}],"paths":{"/api/v1/auth/telegram":{"post":{"summary":"Authenticate user via Telegram Bot/WebApp","description":"Authenticate or register a user using Telegram bot or webapp JSON payload","tags":["Authentication"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TelegramAuthRequest"}}}},"responses":{"200":{"description":"Authentication successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"description":"Invalid Telegram authentication","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"},"example":{"success":false,"message":"Invalid Telegram authentication"}}}},"500":{"$ref":"#/components/responses/ServerError"}}}},"/api/v1/auth/profile":{"get":{"summary":"Get current user profile","description":"Retrieve the authenticated user's profile information","tags":["Authentication"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Profile retrieved successfully","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":true},"data":{"$ref":"#/components/schemas/User"}}}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"},"500":{"$ref":"#/components/responses/ServerError"}}},"put":{"summary":"Update user profile","description":"Update the authenticated user's profile information","tags":["Authentication"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProfileRequest"}}}},"responses":{"200":{"description":"Profile updated successfully","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","example":"Profile updated successfully"},"data":{"$ref":"#/components/schemas/User"}}}}}},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"$ref":"#/components/responses/UnauthorizedError"},"404":{"$ref":"#/components/responses/NotFoundError"},"500":{"$ref":"#/components/responses/ServerError"}}}},"/api/v1/auth/wallet/connect-url":{"get":{"summary":"Generate wallet connection URL with short-lived access token","tags":["Authentication"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"query","name":"redirectUrl","schema":{"type":"string"},"description":"Optional redirect URL after successful connection"}],"responses":{"200":{"description":"Connection URL generated successfully","content":{"application/json":{"schema":{"type":"object","properties":{"url":{"type":"string","description":"The wallet connection URL"},"accessToken":{"type":"string","description":"Short-lived access token"},"expiresAt":{"type":"number","description":"Token expiration timestamp"}}}}}},"401":{"description":"Authentication required"},"500":{"description":"Server error"}}}},"/api/v1/auth/wallet":{"post":{"summary":"Link wallet address to user account","description":"Associate a Stacks wallet address with the authenticated user's account","tags":["Authentication"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LinkWalletRequest"}}}},"responses":{"200":{"description":"Wallet linked successfully","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","example":"Wallet linked successfully"},"data":{"$ref":"#/components/schemas/User"}}}}}},"400":{"description":"Invalid wallet address or already linked","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"},"example":{"success":false,"message":"Wallet address already linked to another account"}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"},"404":{"$ref":"#/components/responses/NotFoundError"},"500":{"$ref":"#/components/responses/ServerError"}}},"delete":{"summary":"Unlink wallet address from user account","description":"Remove the associated wallet address from the authenticated user's account","tags":["Authentication"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Wallet unlinked successfully","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","example":"Wallet unlinked successfully"},"data":{"$ref":"#/components/schemas/User"}}}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"},"404":{"$ref":"#/components/responses/NotFoundError"},"500":{"$ref":"#/components/responses/ServerError"}}}},"/api/v1/balance/plmd":{"get":{"summary":"Get authenticated user's PLMD token balance","description":"Retrieve the authenticated user's PoolMind token balance from their linked wallet","tags":["Balance"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"PLMD balance retrieved successfully","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":true},"data":{"type":"object","properties":{"address":{"type":"string","example":"SP2J6ZY48GV1EZ5V2V5RB9MP66SW86PYKKQ9H6DPR"},"balance":{"type":"number","example":1000000},"balanceFormatted":{"type":"string","example":"1.000000"},"decimals":{"type":"number","example":6}}}}}}}},"400":{"description":"Wallet address required","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Wallet address required"}}}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"},"500":{"description":"Server error","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Failed to fetch PLMD balance"}}}}}}}}},"/api/v1/balance/stx":{"get":{"summary":"Get authenticated user's STX wallet balance","description":"Retrieve the authenticated user's STX balance from their linked wallet","tags":["Balance"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"STX balance retrieved successfully","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":true},"data":{"type":"object","properties":{"address":{"type":"string","example":"SP2J6ZY48GV1EZ5V2V5RB9MP66SW86PYKKQ9H6DPR"},"balance":{"type":"number","example":5000000},"balanceFormatted":{"type":"string","example":"5.000000"},"unit":{"type":"string","example":"microSTX"}}}}}}}},"400":{"description":"Wallet address required","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Wallet address required"}}}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"},"500":{"description":"Server error","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Failed to fetch STX balance"}}}}}}}}},"/api/v1/balance/all":{"get":{"summary":"Get authenticated user's complete balance information","description":"Retrieve both PLMD token balance and STX balance for the authenticated user's linked wallet","tags":["Balance"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"All balances retrieved successfully","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":true},"data":{"type":"object","properties":{"address":{"type":"string","example":"SP2J6ZY48GV1EZ5V2V5RB9MP66SW86PYKKQ9H6DPR"},"plmd":{"type":"object","properties":{"balance":{"type":"number","example":1000000},"balanceFormatted":{"type":"string","example":"1.000000"},"decimals":{"type":"number","example":6}}},"stx":{"type":"object","properties":{"balance":{"type":"number","example":5000000},"balanceFormatted":{"type":"string","example":"5.000000"},"unit":{"type":"string","example":"microSTX"}}}}}}}}}},"400":{"description":"Wallet address required","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Wallet address required"}}}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"},"500":{"description":"Server error","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Failed to fetch balances"}}}}}}}}}}}